Added update version control for settings (using a version.txt file) and the database (using an additional table within the SQL database).
This update constitutes v1.0.3. While changes at this scale are to large
to constitute a patch, as the program is still in Alpha testing, a mere
patch update is represented in the semantic versioning.

The code will automatically push structural changes to all prior
versions, including the new version control requirements. Existing
databases will not require manual changes.